requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-pubsub==0.39.1
google-cloud-storage==1.14.0
google-cloud-translate==1.3.3
google-cloud-vision==0.35.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpc-google-iam-v1==0.11.4
grpcio==1.18.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
nltk==3.4
Pillow==5.4.1
pkg-resources==0.0.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pymongo==3.7.2
pytz==2018.9
requests==2.21.0
rsa==4.0
singledispatch==3.4.0.3
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
